requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
attrs==25.3.0
awscli==1.41.3
boto3==1.39.3
botocore==1.39.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dill==0.4.0
docker==7.1.0
docutils==0.19
exceptiongroup==1.3.0
fastapi==0.115.14
filelock==3.18.0
fonttools==4.58.5
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
h11==0.16.0
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.5.2
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
mock==4.0.3
mpmath==1.3.0
multiprocess==0.70.18
networkx==3.2.1
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.3.0
pathos==0.3.4
pillow==11.3.0
platformdirs==4.3.8
pox==0.3.6
ppft==1.7.7
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rich==14.0.0
rpds-py==0.26.0
rsa==4.7.2
s3transfer==0.13.0
sagemaker==2.247.1
sagemaker-core==1.0.42
schema==0.7.7
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
smdebug-rulesconfig==1.0.1
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tblib==3.1.0
thop==0.1.1.post2209072238
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.162
ultralytics-thop==2.0.14
urllib3==1.26.20
uvicorn==0.35.0
zipp==3.23.0
